Fixes #91.

+++ src/datastructures/huffman_tree/huffman_node.h
@@ -24,7 +24,7 @@ typedef struct huffman_node {
 	enum huffman_node_type type;
 	byte data;
 	uint_least16_t order_no;
-	uint_least32_t weight;
+	uint_least64_t weight;
 	huffman_code *code;
 	uint_least8_t depth;
 	struct huffman_node *parent;
